+++ apps/gperf/src/List_Node.cpp
@@ -1,8 +1,6 @@
 // -*- C++ -*-
 
 /**
- * $Id: List_Node.cpp 93359 2011-02-11 11:33:12Z mcorino $
- *
  * Copyright (C) 1989 Free Software Foundation, Inc.
  * written by Douglas C. Schmidt (schmidt@cs.wustl.edu)
  *
@@ -70,7 +68,9 @@ List_Node::List_Node (char *k, int len)
     length (len),
     slot (0)
 {
-  char *ptr = new char[(option[ALLCHARS] ? len : option.max_keysig_size ()) + 1];
+  char *ptr = new char[(option[ALLCHARS] ?
+                        static_cast<u_int>(len) :
+                        option.max_keysig_size ()) + 1];
   keysig = ptr;
   k[len] = '\0';             // Null terminate KEY to separate it from REST.
